requirements.txt
fastapi
python-multipart
python-dotenv
pandas
flask
flask_cors
mlflow
pydantic
boto3
requests
PyYAML
scikit-learn
xgboost
tenacity
uvicorn
fastapi
python-multipart
python-dotenv
pandas
flask
flask_cors
mlflow == 2.11.3
pydantic
boto3
requests
PyYAML
scikit-learn
xgboost
tenacity
uvicorn
